# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini Ticketing API",
        "description": "Event ticketing API built with Laravel 13, MongoDB and Passport OAuth2.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8004/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List my orders",
                "description": "Returns only the authenticated user's own orders, read from the token. One user can never see another's.",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "responses": {
                    "200": {
                        "description": "A paginated list of your orders"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Place an order",
                "description": "The total is computed server-side from the ticket price. Any total, status or user_id sent by the client is discarded.",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id",
                                    "ticket_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "string",
                                        "example": "6a842696e4e5ed74df0bee52"
                                    },
                                    "ticket_id": {
                                        "type": "string",
                                        "example": "6a84285ecff6bfe1600b0c72"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2,
                                        "maximum": 10,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created and paid"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Not enough tickets, or the event is paused, cancelled, or already past"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}/refund": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Refund an order",
                "description": "Marks the order refunded and returns the tickets to stock, atomically. Only your own orders — someone else's returns 404.",
                "operationId": "0c63ece0d98759426bab78bb5d8ed870",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Changed my mind"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refunded, stock returned"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not your order, or it does not exist"
                    },
                    "422": {
                        "description": "Already refunded, or was never paid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{eventId}/tickets": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Create a ticket type",
                "description": "Adds a ticket type to an event. The event_name is copied from the event, never taken from the request.",
                "operationId": "7c92b31ba000f5ed4c3e920be93b5cd1",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "price",
                                    "quantity_available"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 500
                                    },
                                    "quantity_available": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket type created"
                    },
                    "403": {
                        "description": "Token lacks the events:manage scope"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tickets/{id}": {
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Delete a ticket type",
                "description": "Only possible when no orders reference it.",
                "operationId": "f04259a36f61d75107cf887dac1f5ec8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "403": {
                        "description": "Token lacks the events:manage scope"
                    },
                    "422": {
                        "description": "The ticket type has orders"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Update a ticket type",
                "description": "Partial update. Changing the price does not affect existing orders — they store the price paid at purchase time.",
                "operationId": "a3e53f97dbbb56009c38f140ae83cb31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 600
                                    },
                                    "quantity_available": {
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket type updated"
                    },
                    "403": {
                        "description": "Token lacks the events:manage scope"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Obtain a token from /login, then send it as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}